+++ addOns/openapi/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Warn logs to always include stack trace.
+- Correct generation of empty object.
 
 ## [45] - 2025-03-24
 ### Fixed